backend/requirements.txt
selenium>=4.11.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3<2.0
fastapi>=0.104.0
uvicorn>=0.24.0
